+++ marie/executor/asset_util.py
