+++ marie/executor/extract/__init__.py
@@ -1,10 +1,21 @@
 # isort: skip
 
-from .document_annotator_executor import DocumentAnnotatorExecutor  # isort:skip
-from .document_annotator_embedding_executor import DocumentAnnotatorEmbeddingsExecutor
-from .document_annotator_llm_executor import DocumentAnnotatorLLMExecutor
-from .document_annotator_parser_executor import DocumentAnnotatorParserExecutor
-from .document_annotator_table_llm_executor import DocumentAnnotatorTableLLMExecutor
-from .document_annotator_table_parser_executor import (
+from marie.executor.extract.document_annotator_embedding_executor import (
+    DocumentAnnotatorEmbeddingsExecutor,
+)
+from marie.executor.extract.document_annotator_llm_executor import (
+    DocumentAnnotatorLLMExecutor,
+)
+from marie.executor.extract.document_annotator_parser_executor import (
+    DocumentAnnotatorParserExecutor,
+)
+from marie.executor.extract.document_annotator_table_llm_executor import (
+    DocumentAnnotatorTableLLMExecutor,
+)
+from marie.executor.extract.document_annotator_table_parser_executor import (
     DocumentAnnotatorTableParserExecutor,
+)
+
+from marie.executor.extract.document_annotator_executor import (  # isort:skip
+    DocumentAnnotatorExecutor,
 )
