+++ marie/executor/overlay/__init__.py
@@ -1,1 +1,1 @@
-from .overlay_executor import OverlayExecutor
+from marie.executor.overlay.overlay_executor import OverlayExecutor
