+++ marie/executor/pipeline/__init__.py
@@ -1,4 +1,4 @@
-from .document_llm_pipeline_executor import (
+from marie.executor.pipeline.document_llm_pipeline_executor import (
     DocumentLLMPipelineExecutor,
     PipelineExecutor,
 )
