+++ marie/executor/pipeline/document_pipeline_executor.py
@@ -1,16 +1,16 @@
 import os
 import warnings
-from typing import Optional
+from typing import Any, Optional
 
 import torch
 from docarray import DocList
 
 from marie import requests
 from marie.api.docs import AssetKeyDoc
-from marie.executor.extract.util import create_working_dir
+from marie.executor.asset_util import create_working_dir
 from marie.executor.marie_executor import MarieExecutor
 from marie.executor.mixin import StorageMixin
-from marie.executor.util import get_frames_from_docs, parse_parameters
+from marie.executor.request_util import get_frames_from_docs, parse_parameters
 from marie.logging_core.logger import MarieLogger
 from marie.logging_core.predefined import default_logger as logger
 from marie.models.utils import initialize_device_settings, setup_torch_optimizations
@@ -28,7 +28,8 @@ class PipelineExecutor(MarieExecutor, StorageMixin):
         name: str = "",
         device: Optional[str] = None,
         num_worker_preprocess: int = 4,
-        storage: dict[str, any] = None,
+        storage: dict[str, Any] = None,
+        *args,
         **kwargs,
     ):
         """
