+++ marie/executor/template_matching/__init__.py
@@ -1,1 +1,3 @@
-from .template_matching_executor import TemplateMatchingExecutor
+from marie.executor.template_matching.template_matching_executor import (
+    TemplateMatchingExecutor,
+)
