+++ marie/executor/text/__init__.py
@@ -1,1 +1,4 @@
-from .text_extraction_executor import TextExtractionExecutor, TextExtractionExecutorMock
+from marie.executor.text.text_extraction_executor import (
+    TextExtractionExecutor,
+    TextExtractionExecutorMock,
+)
